pyproject.toml
[tool.poetry]
name = "ycc-hull"
version = "0.2.1"
description = "YCC Hull Backend Service"
authors = ["Lajos Cseppentő <1053931+LajosCseppento@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "~3.9"
# General
aiofiles = "^23.1.0"
pyhumps = "^3.8.0"
# Auth
python-keycloak = "^2.15.3"
# REST API
fastapi = "^0.95.1"
python-multipart = "^0.0.6"
uvicorn = { extras = ["standard"], version = "^0.21.1" }
# DB
oracledb = "^1.3.0"
SQLAlchemy = { extras = ["asyncio", "mypy"], version = "^2.0.9" }

[tool.poetry.dev-dependencies]
black = "^23.3.0"
Faker = "^18.4.0"
flake8 = "^6.0.0"
httpx = "^0.24.0"
mypy = "^1.2.0"
passlib = "^1.7.4"
pylint = "^2.17.2"
pylint-pydantic ="^0.1.8"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-html = "^3.2.0"
pytest-mock = "^3.10.0"
types-aiofiles = "^23.1.0.1"
types-passlib = "^1.7.7.11"

[tool.poetry.scripts]
start = "ycc_hull.main:start"
generate-test-data = "test_data.generator:generate"
regenerate-test-data = "test_data.generator:regenerate"
playground = "test_data.playground:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
